dashboard fix

+++ src/main/java/acme/features/developer/trainingModule/DeveloperTrainingModuleDeleteService.java
@@ -12,12 +12,16 @@
 
 package acme.features.developer.trainingModule;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import acme.client.data.models.Dataset;
 import acme.client.services.AbstractService;
 import acme.entities.training.TrainingModule;
+import acme.entities.training.TrainingSession;
 import acme.roles.Developer;
 
 @Service
@@ -68,8 +72,10 @@ public class DeveloperTrainingModuleDeleteService extends AbstractService<Develo
 	@Override
 	public void perform(final TrainingModule object) {
 		assert object != null;
+		List<TrainingSession> sessions = new ArrayList<>(this.repository.findTrainingSessionsOfTrainingModule(object.getId()));
+		if (sessions.isEmpty())
+			this.repository.delete(object);
 
-		this.repository.delete(object);
 	}
 
 	@Override
